refactor: migrate the indexshelf and linkshelf pairs onto the shared core
Both pairs now carry a ToolSpec, a task table and a thin main(). The index shelf
tools run on a third driver, run_index_main, because their target is a table
file, their log path is built from that table's logical path with no suffix, and
they skip a backup copy inside the log hierarchy so the skip reaches the exit
status; a boolean for that last one is the shrug-flag the data-only ToolSpec rule
forbids. The link shelf tools run on _common.run_main.

With the shared code in it, _shelf_common.py measured 1,827 lines against
deviation (3)'s 1,000-line limit, so it split by family: _shelf_common.py (523)
keeps the versioning helpers and the selection driver, _indexshelf_common.py
(617) and _linkshelf_common.py (712) take one family each.

LOGDIRS and set_log_dirs move from _shelf_common.py to _common.py, beside
log_paths_for, which builds the paths they record: run_main now serves a family
that versions the file it replaces, so both drivers have to be able to record
them.

pds4linkshelf --update no longer raises against an existing shelf. The links
read back from a shelf are the plain tuples that were pickled, not the LinkInfo
objects a fresh scan produces, and generate_links dereferenced .linktext on
both; it now reads either through one accessor, as the pds3 twin's merge does.
The test that pinned the AttributeError is inverted accordingly.
