+++ tests/holdings_maintenance/test_pds3_checksums.py
@@ -167,7 +167,7 @@ def test_reinitialize_versions_the_checksum_file_it_replaces(fresh_tree):
     _shelf_common.move_old() versions the file the task is about to overwrite,
     as <name>_v###.txt beside the run's own log file, one past the highest version
     already there. It reads the shared LOGDIRS list that the run fills in through
-    _shelf_common.set_log_dirs(), so a tool that leaves that list empty versions
+    _common.set_log_dirs(), so a tool that leaves that list empty versions
     nothing; this test is what makes that visible, here and in the pds4 twin.
     """
