+++ tests/holdings_maintenance/test_pds3_linkshelf.py
@@ -157,7 +157,7 @@ def test_update_versions_the_shelf_file_it_replaces(shelved_tree):
     _shelf_common.move_old() versions the shelf the task is about to rewrite, as
     <name>_v###<ext> beside the run's own log file, and copies the `.py` and
     `.pickle` files alongside it. It reads the shared LOGDIRS list that main()
-    fills in through _shelf_common.set_log_dirs(), so a tool that leaves that list
+    fills in through _common.set_log_dirs(), so a tool that leaves that list
     empty versions nothing.
     """
